portfolio-tracker/backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0

# PDF parsing
pdfplumber==0.11.4
PyMuPDF==1.24.10

# AI (Claude Vision fallback for complex PDFs)
anthropic==0.34.2

# HTTP client
httpx==0.27.2

# Data processing
pandas==2.2.3
numpy==2.1.1

# Config / env
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# File uploads
python-multipart==0.0.12

# Redis
redis[asyncio]==5.1.0

# Broker APIs
futu-api==9.4.5408

# Utilities
python-dateutil==2.9.0
